fix(#63): split pure config tests out so the raised timeout covers only AgentRunner tests
Follow-up to the panel's finding (major/gate-relaxed, agreement:1): file-scoped
setDefaultTimeout(15s) also raised the ceiling for the parseAgentSpec and
loadAgentSpecs tests, which do NOT construct AgentRunner — over-broadening the
relaxation and contradicting the comment's 'every test here' claim.

Move those pure config-spec tests to a new agent-specs.test.ts (they belong
there anyway — fast, hermetic, no repo). agent-runner.test.ts now contains ONLY
AgentRunner-against-repo tests, so setDefaultTimeout(15_000) is precisely scoped
to exactly the load-sensitive tests; every other file (including the moved
config tests) keeps bun's 5s fail-fast default. Comment corrected to match.

14 tests unchanged (11 AgentRunner + 3 config), all green; typecheck/lint/format clean.
